requirements.txt
asgiref==3.3.4
Django==3.2.4
django-crispy-forms==1.12.0
django-multiselectfield==0.1.12
numpy==1.20.3
opencv-python==4.5.2.54
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
